Improved delete function, currently logged in user cannot delete itself

+++ app/Http/Controllers/api/v1/UserController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreUser;
 use App\Http\Resources\User as UserResource;
 use App\User;
+use Auth;
 use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -93,6 +94,12 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $currentLoggedInUser = Auth::user();
+
+        if ($user->id === $currentLoggedInUser->id) {
+            return response()->json(['message' => Lang::get('validation.custom.request.400')], 400);
+        }
+
         $delete = $user->delete();
 
         return ($delete === true) ? (response()->json([], 204)) : (response()->json(['message' => Lang::get('validation.custom.request.400')], 400));
